# Conflicts:
#	src/klarna/AdvancedCleanPayment.tsx

+++ src/API/admin.tsx
@@ -1,4 +1,4 @@
-import axios, {AxiosResponse, AxiosError} from 'axios';
+import axios, {AxiosError, AxiosResponse} from 'axios';
 import ConvertTimeSlotToNiceTime from '../components/layout/ConvertTimeSlotToNiceTime';
 import {useState} from 'react';
 
@@ -58,10 +58,21 @@ interface BusinessCustomerData {
 const admin = {
     getAllJobs: async (): Promise<Job[]> => {
         try {
-            const response = await axios.get('http://localhost:8080/api/jobs/getAllJobs');
-            const formattedData = response.data.map((job: Job) => {
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+
+            const response = await axios.get('http://localhost:8080/api/jobs/getAllJobs', {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
+
+            return response.data.map((job: Job) => {
                 const date = new Date(job.date);
-                const formattedDate = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`;
+                const formattedDate = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}
+                -${String(date.getDate()).padStart(2, '0')}`;
                 return {
                     ...job,
                     date: formattedDate,
@@ -70,8 +81,6 @@ const admin = {
                     timeSlot: ConvertTimeSlotToNiceTime(job.timeSlot)
                 };
             });
-            // @ts-ignore
-            return formattedData;
         } catch (error) {
             console.error('Error fetching customer data:', error);
             throw error;
@@ -80,26 +89,30 @@ const admin = {
     getJobByStatus: async (): Promise<Job[]> => {
         try {
             const statuses = ['DONE', 'APPROVED', 'UNAPPROVED', 'PAID', 'CANCELLED'];
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
             const response = await axios.get('http://localhost:8080/api/jobs/getByStatus', {
-                params: {statuses},
+                params: { statuses },
                 paramsSerializer: params => {
                     return `statuses=${statuses.join('&statuses=')}`;
+                },
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
                 }
             });
 
-            const formattedData = response.data.map((job: Job) => {
+            return response.data.map((job: Job) => {
                 const date = new Date(job.date);
                 const formattedDate = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`;
                 return {...job, date: formattedDate};
             });
-
-            return formattedData;
         } catch (error) {
             console.error('Error fetching jobs:', error);
             throw error;
         }
     },
-
     LogIn: (credentials: LoginCredentials): void => {
         const [email, setEmail] = useState<string>('');
         const [password, setPassword] = useState<string>('');
@@ -115,7 +128,7 @@ const admin = {
             axios
                 .post(baseURL, credentials)
                 .then((response: AxiosResponse) => {
-                    // Handle the response here
+
                 })
                 .catch((error: AxiosError) => {
                     console.error('Admin log in failed: ' + error.message);
@@ -123,8 +136,9 @@ const admin = {
                 });
         };
     },
-
     getAvailableEmp: async (date: string, hours: number): Promise<EmployeeData[]> => {
+
+
         if (!date) {
             console.error('Ogiltigt datum.');
             return [];
@@ -135,7 +149,15 @@ const admin = {
                 date: date,
                 lookForAvailableThisManyHours: hours
             };
-            const response = await axios.post(Url, checkEmployees);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, checkEmployees, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log(response.data);
             return response.data;
         } catch (error) {
@@ -143,7 +165,6 @@ const admin = {
             throw error;
         }
     },
-
     createBooking: async (
         jobType: string,
         dateAndTime: string,
@@ -157,7 +178,6 @@ const admin = {
         console.log("................*********" + timeSlotList);
         try {
             const Url = 'http://localhost:8080/api/jobs/createJob';
-
             const bookingData = {
                 jobtype: jobType,
                 date: dateAndTime,
@@ -167,33 +187,50 @@ const admin = {
                 customerId: customerId,
                 message: message
             };
-
-            const response = await axios.post(Url, bookingData);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, bookingData, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log('Booking was created', response.data);
         } catch (error) {
             console.error('Error creating booking', error);
             throw error;
         }
     },
-
     updateJobStatus: async (updateJobDTO: object): Promise<void> => {
         try {
-            const response = await axios.put('http://localhost:8080/api/jobs/updateJob', updateJobDTO);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.put('http://localhost:8080/api/jobs/updateJob', updateJobDTO, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log("update request was made: ", response.status);
         } catch (error) {
             console.error('Error updating job', error);
             throw error;
         }
     },
-
     deleteJob: async (jobId: number): Promise<void> => {
-
         try {
             const Url = `http://localhost:8080/api/jobs/deleteJob`;
             const headers = {
-                'jobId': jobId?.toString()
+                'jobId': jobId?.toString() || '',
+                'Authorization': `Bearer ${sessionStorage.getItem("jwt")}`
             };
-            const response = await axios.delete(Url, {headers});
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.delete(Url, { headers });
             console.log('Deleting job was successful', response.data);
 
         } catch (error) {
@@ -201,33 +238,54 @@ const admin = {
             throw error;
         }
     },
-
     createAdmin: async (adminData: EmployeeData): Promise<void> => {
         try {
             const Url = 'http://localhost:8080/api/employee/createEmployee';
-            const response = await axios.post(Url, adminData);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, adminData, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log('Admin was created', response.data);
         } catch (error) {
             console.error('Error creating admin', error);
             throw error;
         }
     },
-
     createBusinessCustomer: async (businessCustomerData: BusinessCustomerData): Promise<void> => {
         try {
             const Url = 'http://localhost:8080/api/customer/create';
-            const response = await axios.post(Url, businessCustomerData);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, businessCustomerData, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log('Business Customer was created', response.data);
         } catch (error) {
             console.error('Error creating business customer', error);
             throw error;
         }
     },
-
     createEmployee: async (employeeData: EmployeeData): Promise<void> => {
         try {
             const Url = 'http://localhost:8080/api/employee/createEmployee';
-            const response = await axios.post(Url, employeeData);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, employeeData, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log('Employee was created', response.data);
         } catch (error) {
             console.error('Error creating employee', error);
@@ -237,14 +295,21 @@ const admin = {
     createPrivateCustomer: async (privateCustomerData: BusinessCustomerData) => {
         try {
             const Url = 'http://localhost:8080/api/customer/create';
-            const response = await axios.post(Url, privateCustomerData);
+            const jwt = sessionStorage.getItem("jwt");
+            if (!jwt) {
+                throw new Error("JWT not found in sessionStorage");
+            }
+            const response = await axios.post(Url, privateCustomerData, {
+                headers: {
+                    'Authorization': `Bearer ${jwt}`
+                }
+            });
             console.log('Private Customer was created', response.data);
             return response.data;
         } catch (error) {
             console.error('Error creating private customer', error);
             throw error;
         }
-
     },
     getJobDetails: async (jobId: number | null) => {
         try {
@@ -252,8 +317,9 @@ const admin = {
                 const url = `http://localhost:8080/api/jobs/getJob`;
                 const headers = {
                     'jobId': jobId?.toString() || '',
+                    'Authorization': `Bearer ${sessionStorage.getItem("jwt")}`
                 };
-                const response = await axios.get(url, {headers});
+                const response = await axios.get(url, { headers });
                 const data = response.data;
                 console.log(response.data.customerId);
                 if (!data || !data.jobId) {
@@ -275,7 +341,6 @@ const admin = {
             console.error(error);
         }
     },
-
     updateJob: async (
         date: string,
         jobId: number | null,
@@ -290,7 +355,6 @@ const admin = {
         try {
             if (jobId !== null) {
                 const url = `http://localhost:8080/api/jobs/updateJob`;
-
                 const editJobData = {
                     date,
                     jobId,
@@ -301,10 +365,16 @@ const admin = {
                     timeSlot,
                     customerId,
                 };
-
-                console.log(jobId);
-                console.log(customerId);
-                await axios.put(url, editJobData, {params: {message}});
+                const jwt = sessionStorage.getItem("jwt");
+                if (!jwt) {
+                    throw new Error("JWT not found in sessionStorage");
+                }
+                await axios.put(url, editJobData, {
+                    params: { message },
+                    headers: {
+                        'Authorization': `Bearer ${jwt}`
+                    }
+                });
                 console.log('Job was updated');
             }
         } catch (error) {
@@ -315,6 +385,7 @@ const admin = {
 };
 
 export default admin;
+
